+++ etils/enp/__init__.py
@@ -25,6 +25,7 @@ from etils.enp import linalg
 from etils.enp import typing
 from etils.enp.array_spec import ArraySpec
 from etils.enp.array_types import dtypes
+from etils.enp.checking import check_and_normalize_arrays
 from etils.enp.geo_utils import project_onto_plane
 from etils.enp.geo_utils import project_onto_vector
 from etils.enp.interp_utils import interp